fix clang-tidy errors

+++ tests/Unit/NumericalAlgorithms/Interpolation/Test_InterpolationTargetWedgeSectionTorus.cpp
@@ -75,7 +75,7 @@ void test_r_theta_lgl() noexcept {
                                      num_theta)[t]);
         for (size_t p = 0; p < num_phi; ++p) {
           const double phi = 2.0 * M_PI * p / num_phi;
-          const double i = r + t * num_radial + p * num_theta * num_radial;
+          const size_t i = r + t * num_radial + p * num_theta * num_radial;
           get<0>(points)[i] = radius * sin(theta) * cos(phi);
           get<1>(points)[i] = radius * sin(theta) * sin(phi);
           get<2>(points)[i] = radius * cos(theta);
@@ -114,7 +114,7 @@ void test_r_theta_uniform() noexcept {
         const double theta = M_PI * (0.25 + 0.5 * t / (num_theta - 1.0));
         for (size_t p = 0; p < num_phi; ++p) {
           const double phi = 2.0 * M_PI * p / num_phi;
-          const double i = r + t * num_radial + p * num_theta * num_radial;
+          const size_t i = r + t * num_radial + p * num_theta * num_radial;
           get<0>(points)[i] = radius * sin(theta) * cos(phi);
           get<1>(points)[i] = radius * sin(theta) * sin(phi);
           get<2>(points)[i] = radius * cos(theta);
